added remove and rename operations
